+++ espnet_onnx/export/asr/models/encoders/contextual_block_xformer.py
@@ -10,7 +10,7 @@ from espnet.nets.pytorch_backend.transformer.subsampling_without_posenc import \
 
 from espnet_onnx.export.asr.get_config import (get_frontend_config,
                                                get_norm_config)
-from espnet_onnx.export.asr.models.multihead_att import \
+from espnet_onnx.export.asr.models.layers.multihead_att import \
     OnnxMultiHeadedAttention
 from espnet_onnx.utils.abs_model import AbsExportModel
